requirements.txt
streamlit
tensorflow
keras
numpy
opencv-python
pillow_
